pyproject.toml
[tool.poetry]
name = "ez-stocks"
version = "0.1.0"
description = ""
authors = ["jdorado <juancamilo.dorado@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.0"
numpy = "^1.26.0"
yfinance = "^0.2.36"
requests = "^2.31.0"
scipy = "^1.12.0"
statsmodels = "^0.14.1"
matplotlib = "^3.8.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
